AuthX ** version new 1.6.9 ***
